Further refined the logging test.
Also refactored test organisation.
