+++ tests/api/static_content.rs
@@ -1,5 +1,5 @@
+use crate::helpers;
 use crate::helpers::spawn_app;
-use rand::Rng;
 use regex::Regex;
 use std::fs::File;
 use std::io::{BufReader, Read};
@@ -66,8 +66,7 @@ async fn random_url_routes_to_index() {
     let test_app = spawn_app();
     let client = reqwest::Client::new();
     let re = get_index_matching_reg_ex();
-    let urls = get_random_urls(10, 15);
-
+    let urls = helpers::get_random_urls(10, 15);
     // Act
     for url in urls {
         let response = client
@@ -89,7 +88,7 @@ async fn random_url_routes_to_index() {
     }
 }
 
-pub fn get_index_matching_reg_ex() -> Regex {
+fn get_index_matching_reg_ex() -> Regex {
     Regex::new(
         r#"(?m)^<!DOCTYPE html>\r?$
 ^<html lang="en">\r?$
@@ -109,30 +108,4 @@ pub fn get_index_matching_reg_ex() -> Regex {
 "#,
     )
     .expect("Could not parse RegEx.")
-}
-
-fn get_random_urls(size: usize, url_length: u8) -> Vec<String> {
-    let mut res = Vec::new();
-    while res.len() < size {
-        let candidate = get_random_string(url_length);
-        if !candidate.starts_with("/api")
-            && !candidate.starts_with("/login")
-            && !candidate.starts_with('~')
-            && !candidate.chars().next().unwrap().is_ascii_digit()
-        {
-            res.push(candidate);
-        }
-    }
-    res
-}
-
-fn get_random_string(length: u8) -> String {
-    const CHARSET: &[u8] = b"abcdefghijklmnopqrstuvwxyz0123456789~";
-    let mut rng = rand::thread_rng();
-    (0..length)
-        .map(|_| {
-            let index = rng.gen_range(0..CHARSET.len());
-            CHARSET[index] as char
-        })
-        .collect()
 }
